Move languages and i18n scripts to src/utils.
