--- languages.js
@@ -1,9 +1,0 @@
-module.exports = {
-  // TODO: Define the list of languages for this project.
-  languages: {
-    en: 'English',
-    de: 'Deutsch',
-  },
-  // TODO: Define the default language.
-  defaultLanguage: 'en',
-};
