+++ src/layouts/index.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { Header } from '../components/header/header';
 import { StaticNavigation } from '../components/navigation/navigation';
 import { useEffect } from 'react';
-import i18n from '../../i18n';
+import i18n from '../utils/i18n';
 import { CurrentPathProvider } from '../hooks/current_path';
 import { delocalizePath } from '../components/languages/languages';
-import { defaultLanguage } from '../../languages';
+import { defaultLanguage } from '../utils/languages';
 
 declare global {
   interface Window {
