+++ src/utils/languages.ts
@@ -1,0 +1,8 @@
+// TODO: Define the list of languages for this project.
+export const languages = {
+  en: 'English',
+  de: 'Deutsch',
+};
+
+// TODO: Define the default language.
+export const defaultLanguage = 'en';
